backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
redis==5.2.1
arq==0.26.1
sendgrid==6.11.0
httpx==0.28.1
python-multipart==0.0.20
sse-starlette==2.2.1
playwright==1.49.1
anthropic==0.45.2
